# SslSocket::shutdown logs
# VMS-5018 add multicast warning
#CLOUD-922 Added account registration/activation timestamps
#CLOUD-922 Fixed AccountNewTest.account_timestamps test failure because of round error.
#CLOUD-922 Registering cloud account now requires AccountRegistrationData structure, not AccountData which contains several cloud_db generated values
#CLOUD-922 fixed link error
#CLOUD-923 Method naming fixes.
#CLOUD-923 cloud_db: Added system registration timestamp
#CLOUD-941 Added generation of cloudModules.xml based on template
#VMS-5538 Added more logs
#VMS-5565 Added more logs
#VMS-5570 Added assert when invoking null function
Add language selector to lozalised files
Added test validating invited cloud user authenticating in mediaserver
Backed out changeset: 85720bd61866
Bootstrapper - detect language
Close x_vms_4647_tooltips_refactor branch
Cloud promo notification displayed to owners only
Coding style fixes
Coding style fixes.
Enable export to mp4
Enabled -Werror on cloud_connectivity_ut
Enabled -Werror on connection_mediator_ut
First two languages
Fix crash on client closing
Fix header styles
Fix language selector
Fix language.json
Fixed client crash when hovering over bookmarks
Fixed compile error
Fixed crash on test failure
Fixed test crash on failure
Fixed typo
Fixed undefined behaviour in UdpHolePunching.MultiConnect test
Fixed: wrong result code in mediator logs
Fun tests old: disable obsolete tests
Language selector
MOBILE-372: Fix a segfault and a race in MediaPlayer
MOBILE-481: Fix some issues in video aspect ratio handling
MOBILE-483: Improve initial timeline scale
MOBILE-483: Increase default timeline window size to 24 hours
NX_VMS-3522: Wix: Translation fixes: removed &amp and formatted hyperlinks
Project list file for 2.6
QnConnectToCloudWatcher now can use overrided cloud url (needed for tests)
Removed unused code fragment
Removed unused function
Senturian iphone build
Sorting out mess with gtest custom type printer functions.
Spelling fix
Translations strings fix: system -> System
Translations: Cloud Beta made untranslatable
Try to fix 'Bad substitution'
UT-42
Updated Russian translations
Use language list from config.js
Using nx::utils::utcTime() insead of std::chrono::system_clock::now()
VMS-1861 Send statistic after updating system
VMS-2052: Fix resource credentials reading from resourceData.json
VMS-4767 - Make previews on tooltips in notifications the same as in tree
VMS-4852: Add vertical lines to the timeline scrollbar
VMS-5219 - Change text in 'motion on camera' email
VMS-5245: Small fix for Cellinx ptz mapping
VMS-5284 - Get rid of 'thumb' icons
VMS-5337: Add Rest API breaking changes info to web-admin
VMS-5358: 'You have the latest version...' message should not appear until update is finished
VMS-5379 - 'Clear Screen' option for VideoWall doesn't remove icon at User Settings until relogin
VMS-5379 - 'Clear Screen' option for VideoWall doesn't remove icon at User Settings untill relogin
VMS-5385: Fix updates widget lock after cancelling update
VMS-5464: Wrong line height in table in Storage Analytics
VMS-5493: Setup Wizard must be translatable
VMS-5508: Add user's name (or login if name is empty) into tooltip in tree
VMS-5510 - Auto-login to cloud systems
VMS-5510 - Auto-login to cloud systems - cloud host on offline tile
VMS-5512: Screenshot has to be in hi-res
VMS-5513: Alt+Space should invoke the main menu
VMS-5515 - Preloader doesn't show during connecting to System
VMS-5518 - NX client was crashed (build 14105)
VMS-5522 - Add ... for tiles with long owner name
VMS-5525: Bookmark Log. Filter by date doesn't work on first attempt
VMS-5528: There is no any information when PTZ tour is active
VMS-5530: Export Audio in nov doesn't work
VMS-5532 - Fix 'The Word_from_Capital' translation
VMS-5541 - Crash detected in client: f67501de98221a023868bf0e9217e09b
VMS-5542 Axis speakers don't work
VMS-5556 Rtsp stream works in vlc not in Nx
VMS-5557 NX1 crash at stop (build 3.0.0.14114) - libc_do_syscall
VMS-5557 fix
VMS-5559 set ulimit file-max doesn't work
VMS-5567: Vista camera tilt control fix
VMS-5569 - 'Run Application when PC boots' doesn't work
VMS-5573: LIVE button isn't turned ON after unpause camera without archive
VMS-5577 NIC can be disabled during hardware id calculation
VMS-5587: Couldn't bound system to Cloud because /api/cloud_modules.xml is blocked by some other software
VMS-5591:Video layout fix for AV20365
VMS-5594 - Specific web page doesn't open in Client
VMS-5597 Hikvision DS-2CD6362F-I stream profiles are attached to the channels incorrectly
VMS-5600: License failover period should be 30 days
VMS-5601: NX1 mediaserver crash with signal 11, Segmentation fault ( FrameTypeExtractor::getFrameType)
VMS-5602: Live <-> pause flicker when dragging speed slider from 1x to 0x
VMS-5606: After playing last motion chunk, playback doesn't automatically jump at LIVE
VMS-5614: Main menu shows up twice on OS X
generate language.json
hotfix - misspelling
ios signing for senturian
system_clock::time_point gtest printer now prints nanoseconds, not milliseconds
turkish translations-take 2
vms-4405. Backout.
vms-4405. cdb result code UT refactored.
vms-5571. Measure archive cleanup speed.
vms-5589. Don't select storages with free space below 150 mb for recording.
vms-5609. Removed redundant warnings.

<==== Mercurial data ====>
Branch: dev_3.0.0_web

+++ unit_tests/cloud_connectivity_ut/src/cloud_server_socket_ut.cpp
@@ -274,7 +274,7 @@ TEST_F(CloudServerSocketTcpTest, OpenTunnelOnIndication)
     server->moveToListeningState();
 
     // there is no tunnels yet
-    ASSERT_EQ(addressBinder.get(addressManager.key).size(), 0);
+    ASSERT_EQ(addressBinder.get(addressManager.key).size(), 0U);
 
     hpm::api::ConnectionRequestedEvent event;
     event.connectSessionId = String("someSessionId");
@@ -290,7 +290,7 @@ TEST_F(CloudServerSocketTcpTest, OpenTunnelOnIndication)
     std::this_thread::sleep_for(std::chrono::milliseconds(100));
 
     auto list = addressBinder.get(addressManager.key);
-    ASSERT_EQ(list.size(), 1);
+    ASSERT_EQ(list.size(), 1U);
 
     auto client = std::make_unique<TCPSocket>(AF_INET);
     ASSERT_TRUE(client->setNonBlockingMode(true));
@@ -375,7 +375,7 @@ protected:
                             cl_logDEBUG2);
 
                         ASSERT_EQ(code, SystemError::noError);
-                        ASSERT_EQ(size, network::test::kTestMessage.size());
+                        ASSERT_EQ(size, (size_t)network::test::kTestMessage.size());
                     });
 
                 {
